+++ src/Views.Win32/plugin/MupenRRPlugin.cpp
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2026, Mupen64 maintainers, contributors, and original authors (Hacktarux, ShadowPrince, linker).
+ *
+ * SPDX-License-Identifier: GPL-2.0-or-later
+ */
+
+#include "stdafx.h"
+#include <plugin/MupenRRPlugin.hpp>
+
+void MupenRRPlugin::config(HWND hwnd)
+{
+}
+
+void MupenRRPlugin::test(HWND hwnd)
+{
+}
+
+void MupenRRPlugin::about(HWND hwnd)
+{
+}
+
+void MupenRRPlugin::initiate()
+{
+    switch (m_type)
+    {
+    default:
+        RT_ASSERT(false, L"Unsupported plugin type");
+        break;
+    }
+}
+
+void MupenRRPlugin::initiate_dummy()
+{
+}
+
+void MupenRRPlugin::deinitiate_dummy()
+{
+}
