+++ src/Views.Win32/plugin/ZilmarExtPlugin.cpp
@@ -1,0 +1,547 @@
+/*
+ * Copyright (c) 2026, Mupen64 maintainers, contributors, and original authors (Hacktarux, ShadowPrince, linker).
+ *
+ * SPDX-License-Identifier: GPL-2.0-or-later
+ */
+
+// ReSharper disable CppCStyleCast
+
+#include "stdafx.h"
+#include <Config.hpp>
+#include <DialogService.hpp>
+#include <components/MGECompositor.hpp>
+#include <components/Statusbar.hpp>
+#include <plugin/Plugin.hpp>
+#include <plugin/ZilmarExtPlugin.hpp>
+
+#ifndef CALL
+#define CALL _cdecl
+#endif
+
+static size_t ext_fn_config_path(char *data, size_t size)
+{
+    static const std::u8string config_path = IOUtils::config_path().u8string();
+
+    if (data == nullptr) return config_path.size() + 1;
+    if (size < config_path.size() + 1) return 0;
+
+    memcpy(data, config_path.c_str(), config_path.size() + 1);
+    return size + 1;
+}
+
+#pragma region Dummy Functions
+
+static uint32_t CALL dummy_do_rsp_cycles(uint32_t Cycles)
+{
+    return Cycles;
+}
+
+static void CALL dummy_void()
+{
+}
+
+static void CALL dummy_receive_extended_funcs(ZilmarExtSpec::ExtendedFuncs *)
+{
+}
+
+static int32_t CALL dummy_initiate_gfx(ZilmarExtSpec::VideoPluginInfo)
+{
+    return 1;
+}
+
+static int32_t CALL dummy_initiate_audio(ZilmarExtSpec::AudioPluginInfo)
+{
+    return 1;
+}
+
+static void CALL dummy_initiate_controllers(ZilmarExtSpec::InputPluginInfo)
+{
+}
+
+static void CALL dummy_ai_dacrate_changed(int32_t)
+{
+}
+
+static uint32_t CALL dummy_ai_read_length()
+{
+    return 0;
+}
+
+static void CALL dummy_ai_update(int32_t)
+{
+}
+
+static void CALL dummy_controller_command(int32_t, uint8_t *)
+{
+}
+
+static void CALL dummy_get_keys(int32_t, ZilmarExtSpec::Buttons *)
+{
+}
+
+static void CALL dummy_set_keys(int32_t, ZilmarExtSpec::Buttons)
+{
+}
+
+static void CALL dummy_read_controller(int32_t, uint8_t *)
+{
+}
+
+static void CALL dummy_key_down(uint32_t, int32_t)
+{
+}
+
+static void CALL dummy_key_up(uint32_t, int32_t)
+{
+}
+
+static void CALL dummy_initiate_rsp(ZilmarExtSpec::RSPPluginInfo, uint32_t *)
+{
+}
+
+static void CALL dummy_fb_read(uint32_t)
+{
+}
+
+static void CALL dummy_fb_write(uint32_t, uint32_t)
+{
+}
+
+static void CALL dummy_fb_get_framebuffer_info(ZilmarExtSpec::FBInfo *)
+{
+}
+
+static void CALL dummy_move_screen(int32_t, int32_t)
+{
+}
+
+static void CALL dummy_capture_screen(char *)
+{
+    if (!PluginUtil::mge_available())
+    {
+        DialogService::show_dialog(L"The current video plugin doesn't support screenshots.", L"Screenshot", fsvc_error);
+        return;
+    }
+
+    int32_t width{};
+    int32_t height{};
+    MGECompositor::get_video_size(&width, &height);
+
+    std::vector<std::uint8_t> video(width * height * 4);
+    MGECompositor::copy_video(video.data());
+
+    BITMAPINFOHEADER ihdr;
+    ihdr.biSize = sizeof(BITMAPINFOHEADER);
+    ihdr.biWidth = width;
+    ihdr.biHeight = height;
+    ihdr.biPlanes = 1;
+    ihdr.biBitCount = 32;
+    ihdr.biCompression = BI_RGB;
+    ihdr.biSizeImage = width * height * 4;
+    ihdr.biXPelsPerMeter = 0;
+    ihdr.biYPelsPerMeter = 0;
+    ihdr.biClrUsed = 0;
+    ihdr.biClrImportant = 0;
+
+    BITMAPFILEHEADER bhdr;
+    bhdr.bfType = 19778;
+    bhdr.bfSize = sizeof(BITMAPFILEHEADER) + sizeof(BITMAPINFOHEADER) + ihdr.biSizeImage;
+    bhdr.bfReserved1 = bhdr.bfReserved2 = 0;
+    bhdr.bfOffBits = sizeof(BITMAPFILEHEADER) + sizeof(BITMAPINFOHEADER);
+
+    const auto path = Config::screenshot_directory() / std::format("screen{}.bmp", time(nullptr));
+
+    HANDLE hfile;
+    hfile = CreateFile(path.c_str(), GENERIC_WRITE, 0, NULL, CREATE_NEW, FILE_ATTRIBUTE_NORMAL, NULL);
+
+    DWORD written;
+
+    WriteFile(hfile, &bhdr, sizeof(BITMAPFILEHEADER), &written, NULL);
+    WriteFile(hfile, &ihdr, sizeof(BITMAPINFOHEADER), &written, NULL);
+    WriteFile(hfile, video.data(), ihdr.biSizeImage, &written, NULL);
+
+    CloseHandle(hfile);
+}
+
+#pragma endregion
+
+#define GEN_EXTENDED_FUNCS(logger)                                                                                     \
+    ZilmarExtSpec::ExtendedFuncs                                                                                       \
+    {                                                                                                                  \
+        .log_trace = [](const wchar_t *str) { logger->trace(str); },                                                   \
+        .log_info = [](const wchar_t *str) { logger->info(str); },                                                     \
+        .log_warn = [](const wchar_t *str) { logger->warn(str); },                                                     \
+        .log_error = [](const wchar_t *str) { logger->error(str); },                                                   \
+        .get_effective_speed_mode = [](void) { return g_main_ctx.core_ctx->vr_get_effective_speed_mode(); },           \
+        .frame_skipped = [](void) { return g_main_ctx.core_ctx->vr_get_frame_skipped(); },                             \
+        .config_path = ext_fn_config_path, .rcp_counter = g_main_ctx.core_ctx->rcp_counter                             \
+    }
+
+/**
+ * \brief Tries to find the free function exported by the CRT in the specified module.
+ */
+static void (*get_free_function_in_module(HMODULE module))(void *)
+{
+    auto dll_crt_free = (ZilmarExtSpec::DLLCRTFREE)GetProcAddress(module, "DllCrtFree");
+    if (dll_crt_free) return dll_crt_free;
+
+    ULONG size;
+    auto import_descriptor = (PIMAGE_IMPORT_DESCRIPTOR)ImageDirectoryEntryToDataEx(
+        module, true, IMAGE_DIRECTORY_ENTRY_IMPORT, &size, nullptr);
+    if (import_descriptor != nullptr)
+    {
+        while (import_descriptor->Characteristics && import_descriptor->Name)
+        {
+            auto importDllName = (LPCSTR)((PBYTE)module + import_descriptor->Name);
+            auto importDllHandle = GetModuleHandleA(importDllName);
+            if (importDllHandle != nullptr)
+            {
+                dll_crt_free = (ZilmarExtSpec::DLLCRTFREE)GetProcAddress(importDllHandle, "free");
+                if (dll_crt_free != nullptr) return dll_crt_free;
+            }
+
+            import_descriptor++;
+        }
+    }
+
+    return free;
+}
+
+void ZilmarExtPlugin::config(HWND hwnd)
+{
+    initiate_dummy();
+
+    const auto dll_config = (ZilmarExtSpec::DLLCONFIG)GetProcAddress(m_module, "DllConfig");
+
+    if (dll_config)
+        dll_config(hwnd);
+    else
+    {
+        DialogService::show_dialog(
+            std::format(L"'{}' has no configuration.", IOUtils::to_wide_string(this->name())).c_str(), L"Plugin",
+            fsvc_error, hwnd);
+    }
+
+    deinitiate_dummy();
+}
+
+void ZilmarExtPlugin::test(HWND hwnd)
+{
+    initiate_dummy();
+    const auto dll_test = (ZilmarExtSpec::DLLTEST)GetProcAddress(m_module, "DllTest");
+    if (dll_test) dll_test(hwnd);
+    deinitiate_dummy();
+}
+
+void ZilmarExtPlugin::about(HWND hwnd)
+{
+    initiate_dummy();
+    const auto dll_about = (ZilmarExtSpec::DLLABOUT)GetProcAddress(m_module, "DllAbout");
+    if (dll_about) dll_about(hwnd);
+    deinitiate_dummy();
+}
+
+void ZilmarExtPlugin::initiate()
+{
+    switch (m_type)
+    {
+    case ZilmarExtSpec::PluginType::Video: {
+        g_view_logger->trace("Initiating video plugin...");
+        ZilmarExtSpec::INITIATEGFX initiate_gfx{};
+
+        FUNC(g_plugin_funcs.video_change_window, ZilmarExtSpec::CHANGEWINDOW, dummy_void, "ChangeWindow");
+        FUNC(g_plugin_funcs.video_close_dll, ZilmarExtSpec::CLOSEDLL, dummy_void, "CloseDLL");
+        FUNC(initiate_gfx, ZilmarExtSpec::INITIATEGFX, dummy_initiate_gfx, "InitiateGFX");
+        FUNC(g_plugin_funcs.video_process_dlist, ZilmarExtSpec::PROCESSDLIST, dummy_void, "ProcessDList");
+        FUNC(g_plugin_funcs.video_process_rdp_list, ZilmarExtSpec::PROCESSRDPLIST, dummy_void, "ProcessRDPList");
+        FUNC(g_plugin_funcs.video_rom_closed, ZilmarExtSpec::ROMCLOSED, dummy_void, "RomClosed");
+        FUNC(g_plugin_funcs.video_rom_open, ZilmarExtSpec::ROMOPEN, dummy_void, "RomOpen");
+        FUNC(g_plugin_funcs.video_show_cfb, ZilmarExtSpec::SHOWCFB, dummy_void, "ShowCFB");
+        FUNC(g_plugin_funcs.video_update_screen, ZilmarExtSpec::UPDATESCREEN, dummy_void, "UpdateScreen");
+        FUNC(g_plugin_funcs.video_vi_status_changed, ZilmarExtSpec::VISTATUSCHANGED, dummy_void, "ViStatusChanged");
+        FUNC(g_plugin_funcs.video_vi_width_changed, ZilmarExtSpec::VIWIDTHCHANGED, dummy_void, "ViWidthChanged");
+        FUNC(g_plugin_funcs.video_move_screen, ZilmarExtSpec::MOVESCREEN, dummy_move_screen, "MoveScreen");
+        FUNC(g_plugin_funcs.video_capture_screen, ZilmarExtSpec::CAPTURESCREEN, dummy_capture_screen, "CaptureScreen");
+        FUNC(g_plugin_funcs.video_read_screen, ZilmarExtSpec::READSCREEN,
+             (ZilmarExtSpec::READSCREEN)GetProcAddress(m_module, "ReadScreen2"), "ReadScreen");
+        FUNC(g_plugin_funcs.video_get_video_size, ZilmarExtSpec::GETVIDEOSIZE, nullptr, "mge_get_video_size");
+        FUNC(g_plugin_funcs.video_read_video, ZilmarExtSpec::READVIDEO, nullptr, "mge_read_video2");
+        FUNC(g_plugin_funcs.video_fb_read, ZilmarExtSpec::FBREAD, dummy_fb_read, "FBRead");
+        FUNC(g_plugin_funcs.video_fb_write, ZilmarExtSpec::FBWRITE, dummy_fb_write, "FBWrite");
+        FUNC(g_plugin_funcs.video_fb_get_frame_buffer_info, ZilmarExtSpec::FBGETFRAMEBUFFERINFO,
+             dummy_fb_get_framebuffer_info, "FBGetFrameBufferInfo");
+        g_plugin_funcs.video_dll_crt_free = get_free_function_in_module(m_module);
+
+        gfx_info.main_hwnd = g_main_ctx.hwnd;
+        gfx_info.statusbar_hwnd = g_config.is_statusbar_enabled ? Statusbar::hwnd() : nullptr;
+        gfx_info.byteswapped = 1;
+        gfx_info.rom = g_main_ctx.core_ctx->rom;
+        gfx_info.rdram = (uint8_t *)g_main_ctx.core_ctx->rdram;
+        gfx_info.dmem = (uint8_t *)g_main_ctx.core_ctx->SP_DMEM;
+        gfx_info.imem = (uint8_t *)g_main_ctx.core_ctx->SP_IMEM;
+        gfx_info.mi_intr_reg = &g_main_ctx.core_ctx->MI_register->mi_intr_reg;
+        gfx_info.dpc_start_reg = &g_main_ctx.core_ctx->dpc_register->dpc_start;
+        gfx_info.dpc_end_reg = &g_main_ctx.core_ctx->dpc_register->dpc_end;
+        gfx_info.dpc_current_reg = &g_main_ctx.core_ctx->dpc_register->dpc_current;
+        gfx_info.dpc_status_reg = &g_main_ctx.core_ctx->dpc_register->dpc_status;
+        gfx_info.dpc_clock_reg = &g_main_ctx.core_ctx->dpc_register->dpc_clock;
+        gfx_info.dpc_bufbusy_reg = &g_main_ctx.core_ctx->dpc_register->dpc_bufbusy;
+        gfx_info.dpc_pipebusy_reg = &g_main_ctx.core_ctx->dpc_register->dpc_pipebusy;
+        gfx_info.dpc_tmem_reg = &g_main_ctx.core_ctx->dpc_register->dpc_tmem;
+        gfx_info.vi_status_reg = &g_main_ctx.core_ctx->vi_register->vi_status;
+        gfx_info.vi_origin_reg = &g_main_ctx.core_ctx->vi_register->vi_origin;
+        gfx_info.vi_width_reg = &g_main_ctx.core_ctx->vi_register->vi_width;
+        gfx_info.vi_intr_reg = &g_main_ctx.core_ctx->vi_register->vi_v_intr;
+        gfx_info.vi_v_current_line_reg = &g_main_ctx.core_ctx->vi_register->vi_current;
+        gfx_info.vi_timing_reg = &g_main_ctx.core_ctx->vi_register->vi_burst;
+        gfx_info.vi_v_sync_reg = &g_main_ctx.core_ctx->vi_register->vi_v_sync;
+        gfx_info.vi_h_sync_reg = &g_main_ctx.core_ctx->vi_register->vi_h_sync;
+        gfx_info.vi_leap_reg = &g_main_ctx.core_ctx->vi_register->vi_leap;
+        gfx_info.vi_h_start_reg = &g_main_ctx.core_ctx->vi_register->vi_h_start;
+        gfx_info.vi_v_start_reg = &g_main_ctx.core_ctx->vi_register->vi_v_start;
+        gfx_info.vi_v_burst_reg = &g_main_ctx.core_ctx->vi_register->vi_v_burst;
+        gfx_info.vi_x_scale_reg = &g_main_ctx.core_ctx->vi_register->vi_x_scale;
+        gfx_info.vi_y_scale_reg = &g_main_ctx.core_ctx->vi_register->vi_y_scale;
+        gfx_info.check_interrupts = dummy_void;
+
+        g_plugin_funcs.video_extended_funcs = GEN_EXTENDED_FUNCS(g_video_logger);
+        gfx_info.extended_funcs = &g_plugin_funcs.video_extended_funcs;
+        initiate_gfx(gfx_info);
+
+        bool compat_error = false;
+
+        if (!g_plugin_funcs.video_read_video && GetProcAddress(m_module, "mge_read_video")) compat_error = true;
+
+        if (compat_error)
+        {
+            const auto msg =
+                std::format(L"The plugin {} is incompatible with this version of Mupen64 and may not work properly.",
+                            IOUtils::to_wide_string(m_name));
+            DialogService::show_dialog(msg.c_str(), L"Plugin Incompatibility", fsvc_error);
+        }
+
+        break;
+    }
+    case ZilmarExtSpec::PluginType::Audio: {
+        g_view_logger->trace("Initiating audio plugin...");
+        ZilmarExtSpec::INITIATEAUDIO initiate_audio{};
+
+        FUNC(g_plugin_funcs.audio_close_dll_audio, ZilmarExtSpec::CLOSEDLL, dummy_void, "CloseDLL");
+        FUNC(g_plugin_funcs.audio_ai_dacrate_changed, ZilmarExtSpec::AIDACRATECHANGED, dummy_ai_dacrate_changed,
+             "AiDacrateChanged");
+        FUNC(g_plugin_funcs.audio_ai_len_changed, ZilmarExtSpec::AILENCHANGED, dummy_void, "AiLenChanged");
+        FUNC(g_plugin_funcs.audio_ai_read_length, ZilmarExtSpec::AIREADLENGTH, dummy_ai_read_length, "AiReadLength");
+        FUNC(initiate_audio, ZilmarExtSpec::INITIATEAUDIO, dummy_initiate_audio, "InitiateAudio");
+        FUNC(g_plugin_funcs.audio_rom_closed, ZilmarExtSpec::ROMCLOSED, dummy_void, "RomClosed");
+        FUNC(g_plugin_funcs.audio_rom_open, ZilmarExtSpec::ROMOPEN, dummy_void, "RomOpen");
+        FUNC(g_plugin_funcs.audio_process_alist, ZilmarExtSpec::PROCESSALIST, dummy_void, "ProcessAList");
+        FUNC(g_plugin_funcs.audio_ai_update, ZilmarExtSpec::AIUPDATE, dummy_ai_update, "AiUpdate");
+
+        audio_info.main_hwnd = g_main_ctx.hwnd;
+        audio_info.hinst = g_main_ctx.hinst;
+        audio_info.byteswapped = 1;
+        audio_info.rom = g_main_ctx.core_ctx->rom;
+        audio_info.rdram = (uint8_t *)g_main_ctx.core_ctx->rdram;
+        audio_info.dmem = (uint8_t *)g_main_ctx.core_ctx->SP_DMEM;
+        audio_info.imem = (uint8_t *)g_main_ctx.core_ctx->SP_IMEM;
+        audio_info.mi_intr_reg = &dummy_dw;
+        audio_info.ai_dram_addr_reg = &g_main_ctx.core_ctx->ai_register->ai_dram_addr;
+        audio_info.ai_len_reg = &g_main_ctx.core_ctx->ai_register->ai_len;
+        audio_info.ai_control_reg = &g_main_ctx.core_ctx->ai_register->ai_control;
+        audio_info.ai_status_reg = &dummy_dw;
+        audio_info.ai_dacrate_reg = &g_main_ctx.core_ctx->ai_register->ai_dacrate;
+        audio_info.ai_bitrate_reg = &g_main_ctx.core_ctx->ai_register->ai_bitrate;
+
+        audio_info.check_interrupts = dummy_void;
+
+        g_plugin_funcs.audio_extended_funcs = GEN_EXTENDED_FUNCS(g_audio_logger);
+        audio_info.extended_funcs = &g_plugin_funcs.audio_extended_funcs;
+        initiate_audio(audio_info);
+        break;
+    }
+    case ZilmarExtSpec::PluginType::Input: {
+        g_view_logger->trace("Initiating input plugin...");
+
+        ZilmarExtSpec::OLD_INITIATECONTROLLERS old_initiate_controllers{};
+        ZilmarExtSpec::INITIATECONTROLLERS initiate_controllers{};
+
+        FUNC(g_plugin_funcs.input_close_dll, ZilmarExtSpec::CLOSEDLL, dummy_void, "CloseDLL");
+        FUNC(g_plugin_funcs.input_controller_command, ZilmarExtSpec::CONTROLLERCOMMAND, dummy_controller_command,
+             "ControllerCommand");
+        FUNC(g_plugin_funcs.input_get_keys, ZilmarExtSpec::GETKEYS, dummy_get_keys, "GetKeys");
+        FUNC(g_plugin_funcs.input_set_keys, ZilmarExtSpec::SETKEYS, dummy_set_keys, "SetKeys");
+        if (m_version == 0x0101)
+        {
+            FUNC(initiate_controllers, ZilmarExtSpec::INITIATECONTROLLERS, dummy_initiate_controllers,
+                 "InitiateControllers");
+        }
+        else
+        {
+            FUNC(old_initiate_controllers, ZilmarExtSpec::OLD_INITIATECONTROLLERS, nullptr, "InitiateControllers");
+        }
+        FUNC(g_plugin_funcs.input_read_controller, ZilmarExtSpec::READCONTROLLER, dummy_read_controller,
+             "ReadController");
+        FUNC(g_plugin_funcs.input_rom_closed, ZilmarExtSpec::ROMCLOSED, dummy_void, "RomClosed");
+        FUNC(g_plugin_funcs.input_rom_open, ZilmarExtSpec::ROMOPEN, dummy_void, "RomOpen");
+        FUNC(g_plugin_funcs.input_key_down, ZilmarExtSpec::KEYDOWN, dummy_key_down, "WM_KeyDown");
+        FUNC(g_plugin_funcs.input_key_up, ZilmarExtSpec::KEYUP, dummy_key_up, "WM_KeyUp");
+
+        control_info.main_hwnd = g_main_ctx.hwnd;
+        control_info.hinst = g_main_ctx.hinst;
+        control_info.byteswapped = 1;
+        control_info.header = g_main_ctx.core_ctx->rom;
+
+        std::array<ZilmarExtSpec::Controller, 4> tmp_controllers{};
+        control_info.controllers = tmp_controllers.data();
+
+        g_plugin_funcs.input_extended_funcs = GEN_EXTENDED_FUNCS(g_input_logger);
+        control_info.extended_funcs = &g_plugin_funcs.input_extended_funcs;
+
+        if (m_version == 0x0101)
+            initiate_controllers(control_info);
+        else
+            old_initiate_controllers(g_main_ctx.hwnd, tmp_controllers.data());
+
+        for (size_t i = 0; i < std::size(tmp_controllers); ++i)
+        {
+            g_main_ctx.core.controls[i] = tmp_controllers[i].to_core_controller();
+        }
+        break;
+    }
+    case ZilmarExtSpec::PluginType::RSP: {
+        g_view_logger->trace("Initiating RSP plugin...");
+        ZilmarExtSpec::INITIATERSP initiate_rsp{};
+
+        FUNC(g_plugin_funcs.rsp_close_dll, ZilmarExtSpec::CLOSEDLL, dummy_void, "CloseDLL");
+        FUNC(g_plugin_funcs.rsp_do_rsp_cycles, ZilmarExtSpec::DORSPCYCLES, dummy_do_rsp_cycles, "DoRspCycles");
+        FUNC(initiate_rsp, ZilmarExtSpec::INITIATERSP, dummy_initiate_rsp, "InitiateRSP");
+        FUNC(g_plugin_funcs.rsp_rom_closed, ZilmarExtSpec::ROMCLOSED, dummy_void, "RomClosed");
+
+        rsp_info.byteswapped = 1;
+        rsp_info.rdram = (uint8_t *)g_main_ctx.core_ctx->rdram;
+        rsp_info.dmem = (uint8_t *)g_main_ctx.core_ctx->SP_DMEM;
+        rsp_info.imem = (uint8_t *)g_main_ctx.core_ctx->SP_IMEM;
+        rsp_info.mi_intr_reg = &g_main_ctx.core_ctx->MI_register->mi_intr_reg;
+        rsp_info.sp_mem_addr_reg = &g_main_ctx.core_ctx->sp_register->sp_mem_addr_reg;
+        rsp_info.sp_dram_addr_reg = &g_main_ctx.core_ctx->sp_register->sp_dram_addr_reg;
+        rsp_info.sp_rd_len_reg = &g_main_ctx.core_ctx->sp_register->sp_rd_len_reg;
+        rsp_info.sp_wr_len_reg = &g_main_ctx.core_ctx->sp_register->sp_wr_len_reg;
+        rsp_info.sp_status_reg = &g_main_ctx.core_ctx->sp_register->sp_status_reg;
+        rsp_info.sp_dma_full_reg = &g_main_ctx.core_ctx->sp_register->sp_dma_full_reg;
+        rsp_info.sp_dma_busy_reg = &g_main_ctx.core_ctx->sp_register->sp_dma_busy_reg;
+        rsp_info.sp_pc_reg = &g_main_ctx.core_ctx->rsp_register->rsp_pc;
+        rsp_info.sp_semaphore_reg = &g_main_ctx.core_ctx->sp_register->sp_semaphore_reg;
+        rsp_info.dpc_start_reg = &g_main_ctx.core_ctx->dpc_register->dpc_start;
+        rsp_info.dpc_end_reg = &g_main_ctx.core_ctx->dpc_register->dpc_end;
+        rsp_info.dpc_current_reg = &g_main_ctx.core_ctx->dpc_register->dpc_current;
+        rsp_info.dpc_status_reg = &g_main_ctx.core_ctx->dpc_register->dpc_status;
+        rsp_info.dpc_clock_reg = &g_main_ctx.core_ctx->dpc_register->dpc_clock;
+        rsp_info.dpc_bufbusy_reg = &g_main_ctx.core_ctx->dpc_register->dpc_bufbusy;
+        rsp_info.dpc_pipebusy_reg = &g_main_ctx.core_ctx->dpc_register->dpc_pipebusy;
+        rsp_info.dpc_tmem_reg = &g_main_ctx.core_ctx->dpc_register->dpc_tmem;
+        rsp_info.check_interrupts = dummy_void;
+        rsp_info.process_dlist_list = g_plugin_funcs.video_process_dlist;
+        rsp_info.process_alist_list = g_plugin_funcs.audio_process_alist;
+        rsp_info.process_rdp_list = g_plugin_funcs.video_process_rdp_list;
+        rsp_info.show_cfb = g_plugin_funcs.video_show_cfb;
+
+        g_plugin_funcs.rsp_extended_funcs = GEN_EXTENDED_FUNCS(g_rsp_logger);
+        rsp_info.extended_funcs = &g_plugin_funcs.rsp_extended_funcs;
+
+        int32_t i = 4;
+        initiate_rsp(rsp_info, (uint32_t *)&i);
+        break;
+    }
+    }
+}
+
+void ZilmarExtPlugin::initiate_dummy()
+{
+    Main::init_sdl();
+
+    if (m_spec != Spec::ZilmarExt) return;
+
+    switch (m_type)
+    {
+    case ZilmarExtSpec::PluginType::Video: {
+        if (!g_main_ctx.core_ctx->vr_get_launched())
+        {
+            // NOTE: Since olden days, dummy render target hwnd was the statusbar.
+            dummy_video_info.main_hwnd = Statusbar::hwnd();
+            dummy_video_info.statusbar_hwnd = Statusbar::hwnd();
+
+            g_plugin_funcs.video_extended_funcs = GEN_EXTENDED_FUNCS(g_video_logger);
+            dummy_video_info.extended_funcs = &g_plugin_funcs.video_extended_funcs;
+
+            const auto initiate_gfx = (ZilmarExtSpec::INITIATEGFX)GetProcAddress(m_module, "InitiateGFX");
+            if (initiate_gfx && !initiate_gfx(dummy_video_info))
+            {
+                DialogService::show_dialog(L"Couldn't initialize video plugin.", L"Core", fsvc_information);
+            }
+        }
+
+        break;
+    }
+    case ZilmarExtSpec::PluginType::Audio: {
+        if (!g_main_ctx.core_ctx->vr_get_launched())
+        {
+            g_plugin_funcs.audio_extended_funcs = GEN_EXTENDED_FUNCS(g_audio_logger);
+            dummy_audio_info.extended_funcs = &g_plugin_funcs.audio_extended_funcs;
+
+            const auto initiate_audio = (ZilmarExtSpec::INITIATEAUDIO)GetProcAddress(m_module, "InitiateAudio");
+            if (initiate_audio && !initiate_audio(dummy_audio_info))
+            {
+                DialogService::show_dialog(L"Couldn't initialize audio plugin.", L"Core", fsvc_information);
+            }
+        }
+
+        break;
+    }
+    case ZilmarExtSpec::PluginType::Input: {
+        if (!g_main_ctx.core_ctx->vr_get_launched())
+        {
+            g_plugin_funcs.input_extended_funcs = GEN_EXTENDED_FUNCS(g_input_logger);
+            dummy_control_info.extended_funcs = &g_plugin_funcs.input_extended_funcs;
+
+            if (m_version == 0x0101)
+            {
+                const auto initiate_controllers =
+                    (ZilmarExtSpec::INITIATECONTROLLERS)GetProcAddress(m_module, "InitiateControllers");
+                if (initiate_controllers) initiate_controllers(dummy_control_info);
+            }
+            else
+            {
+                const auto old_initiate_controllers =
+                    (ZilmarExtSpec::OLD_INITIATECONTROLLERS)GetProcAddress(m_module, "InitiateControllers");
+
+                std::fill(std::begin(dummy_controllers), std::end(dummy_controllers), ZilmarExtSpec::Controller{});
+                if (old_initiate_controllers) old_initiate_controllers(g_main_ctx.hwnd, dummy_controllers);
+            }
+        }
+
+        break;
+    }
+    case ZilmarExtSpec::PluginType::RSP: {
+        if (!g_main_ctx.core_ctx->vr_get_launched())
+        {
+            g_plugin_funcs.rsp_extended_funcs = GEN_EXTENDED_FUNCS(g_rsp_logger);
+            dummy_rsp_info.extended_funcs = &g_plugin_funcs.rsp_extended_funcs;
+
+            auto initiateRSP = (ZilmarExtSpec::INITIATERSP)GetProcAddress(m_module, "InitiateRSP");
+            uint32_t i = 0;
+            if (initiateRSP) initiateRSP(dummy_rsp_info, &i);
+        }
+
+        break;
+    }
+    default:
+        RT_ASSERT(false, L"Unknown plugin type");
+    }
+}
+
+void ZilmarExtPlugin::deinitiate_dummy()
+{
+    if (g_main_ctx.core_ctx->vr_get_launched()) return;
+    const auto close_dll = (ZilmarExtSpec::CLOSEDLL)GetProcAddress(m_module, "CloseDLL");
+    if (close_dll) close_dll();
+}
